+++ apps/api-admin/src/main/java/com/ott/api_admin/content/vo/IngestJobResult.java
@@ -1,0 +1,12 @@
+package com.ott.api_admin.content.vo;
+
+import com.ott.domain.common.MediaType;
+
+public record IngestJobResult(
+        Long mediaId,
+        Long ingestJobId,
+        String originObjectKey,
+        Long fileSize,
+        MediaType mediaType
+) {
+}
