+++ modules/domain/src/main/java/com/ott/domain/ingest_job/repository/IngestJobRepository.java
@@ -4,4 +4,6 @@ import com.ott.domain.ingest_job.domain.IngestJob;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IngestJobRepository extends JpaRepository<IngestJob, Long>, IngestJobRepositoryCustom {
+
+    boolean existsByMediaId(Long mediaId);
 }
